created new manifest service
