+++ application/backend/tests/service-tests/releases.manifest-helper.test.ts
@@ -1,7 +1,7 @@
 import { beforeEachCommon } from "./releases.common";
 import { registerTypes } from "./setup";
 import { Client } from "edgedb";
-import { createReleaseManifest } from "../../src/business/services/_release-manifest-helper";
+import { createReleaseManifest } from "../../src/business/services/manifests/_manifest-helper";
 import { THOUSAND_GENOMES_SYSTEM } from "../../src/test-data/insert-test-data-10f-helpers";
 
 let edgeDbClient: Client;
